requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.1
scikit-learn==1.4.1.post1
numpy==1.26.4
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
aiofiles==23.2.1
httpx==0.27.0
